+++ e2e/tests/plugins.e2e.js
@@ -513,7 +513,7 @@ describe('plugins', () => {
       ...usePlugins(plugins.all),
       ...usePluginsConfig({
         LINK: {
-          linkPanelAddons: ['anchor'],
+          linkTypes: { anchor: true },
         },
       }),
     };