+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -451,7 +451,7 @@ const config = {
     //   },
     // },
     onClick: (event, url) => console.log('link clicked!', url),
-    linkPanelAddons: ['anchor'],
+    linkTypes: { anchor: true },
   },
   [SOUND_CLOUD_TYPE]: {
     // toolbar: {