+++ examples/storybook/stories/Plugins/Anchor/AnchorPluginStory.js
@@ -13,7 +13,11 @@ import ViewerWrapper from '../../Components/ViewerWrapper';
 export default () => {
   return (
     <Page title="Anchor Plugin">
-      <h4>link config is needed: linkPanelAddons: [`anchor`]</h4>
+      <h4>enable anchors from editor plugins configuration </h4>
+      <h4>
+        add to link config{' '}
+        <code style={{ backgroundColor: 'lightgray' }}>linkTypes: {'{ anchor: true }'}</code>
+      </h4>
       <h5>As of today, anchor can be linked to one of the following:</h5>
       <h5>
         regular text (paragraphs), headers, code blocks, quotes, images, galleries, videos, maps,