+++ packages/editor-common/web/src/Base/buttons/BlockLinkButton.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import EditorModals from '../../Modals/EditorModals';
 import { getModalStyles } from '../../Utils/getModalStyles';
 import LinkButton from '../../Components/LinkComponents/LinkButton';
+import { isEmpty } from 'lodash';
 
 //Atomic Blocks Link Button
 class BlockLinkButton extends Component {
@@ -28,10 +29,14 @@ class BlockLinkButton extends Component {
       innerModal,
       toolbarOffsetTop,
       toolbarOffsetLeft,
-      linkPanelAddons,
+      linkTypes,
       editorState,
     } = this.props;
-    const OriginalLinkPanel = !linkPanelAddons || linkPanelAddons.length === 0 || unchangedUrl;
+    const OriginalLinkPanel =
+      !linkTypes ||
+      isEmpty(linkTypes) ||
+      !Object.values(linkTypes).find(addon => !!addon) ||
+      unchangedUrl;
     const modalStyles = getModalStyles({ fullScreen: !OriginalLinkPanel, isMobile });
     const commonPanelProps = {
       componentState,
@@ -44,7 +49,7 @@ class BlockLinkButton extends Component {
       modalName: EditorModals.BLOCK_LINK_MODAL,
       uiSettings,
       unchangedUrl,
-      linkPanelAddons,
+      linkTypes,
       editorState,
     };
     if (isMobile) {
@@ -108,7 +113,7 @@ BlockLinkButton.propTypes = {
   innerModal: PropTypes.object,
   toolbarOffsetTop: PropTypes.string,
   toolbarOffsetLeft: PropTypes.string,
-  linkPanelAddons: PropTypes.array,
+  linkTypes: PropTypes.object,
   editorState: PropTypes.object,
 };