+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ### :bug: Bug Fix
 - `plugin-link`
   - [#1409](https://github.com/wix-incubator/rich-content/pull/1409) link viewer without onClick from settings
+- `plugin-link`
+  - [#1402](https://github.com/wix-incubator/rich-content/pull/1402) anchor - config changes, ricos documentation, schema fix
 </details>
 <hr/>